提高覆盖率
